Issue 133: GZIP File Archive Java Delegate.
This adds support for `.gz`, `.tar.gz`, `.bz2`, and `.tar.bz2`.
Additional formats are possible but require extra work that is left for later.

The task `CompressFileTask` accepts three parameters:
- source: the source file or directory.
- destination: the destination directory (not the destination file).

When the source is a file, it will be compressed (`example.txt` would become `example.txt.gz` for a `gz` format).
When the source is a directory, the directory will be tarballed and then compressed (`examples/` would become `examples.tar.gz` for a `gz` format).