Issue 113: Change Module and Deployment Descriptors to be consistent with standard FOLIO practices.
Relocate the the Descriptors.
The `descriptors/` directory should be at the base project root but in this specific case with service as a sub-directory the `descriptors/` directory must be under `service/`.

Rename the Descriptors to be a template and bring in the template processing plugins into the POM file.

Change the variable behavior in the Descriptors to be consistent with how standard FOLIO practices use them (`${project.version}` instead of `@project.version@` for example).
This module is again an exceptional case where the project is under `service/`.
The standard `${artifactId}` cannot be used because that ends up being `workflow-service` rather than `mod-workflow`.
Therefore, hard code `mod-workflow` in places where `mod-workflow` must be used instead of `workflow-service`.

Added `-XX:MaxRAMPercentage=66.0` for consistency with other projects.
Some projects use a value like 85.0.
I don't have a realistic clue here on what value to use and I am open to not having this environment variable set at all if a proposal to remove it is given.

The Deployment Descriptor used to have only `--server.port=%p`.
Other FOLIO projects have both `-Dport=%p` and `-Dhttp.port=%p`.
This plays it safe by having all three.