Issue 113: Remove shade plugin due to inherent problems and bugs.
Errors regarding multiple registrations of `AttributeConverter` have surfaced.
This is a bug between the shade plugin and hibernate as far as I can tell from my online research.

I believe that this is sufficient justification to deviate from the Folio practice of using the shade plugin.
Strip out the shade plugin entirely and update the `Dockerfile`.

I also noticed that the default `EXPOSE` for this project should be `9001` and not `9000`.