Issue 113: Prepend 'workflow.' to each permission.
This gets the permissions structure to be compliant by having the module name as part of the permission name.
Without this, potential cross-module conflicts exist.

This has a side-effect of `workflow.workflow.` as a permission.
The additional workflow represents the `workflow` path.
This is not removed to avoid potential naming conflicts as well.