MODWRKFLOW-2: Enable use of Lombok and extract all methods into interfaces.
Lombok is added and is in use for all of the models.

The methods extracted into interfaces have their interfaces prefixed with `Has` and are under the `has` directory.
There is a `has/common` directory for interfaces that are specific to a given model.

The methods that are found to be used in multiple models have been provided in their own `Has` interface.
I attempted to break out the `Has` interfaces into groups of purpose, such as `HasVersioning`.
These could then support multiple methods that would be always grouped together.
The grouping of everything and how they should be group is not entirely clear at this point and so I opted to have most of the `Has` class be single purpose.
The `Has` `Common` interfaces are catch-all for a given model that is not broken out.
These may be changed in the future once the design is better hashed out.
Hashing this it out is considered out of scope.

There are `version` and `versionTag` properties.
It seems like these could be the same thing.
The `versionTag` is Camunda-specific.
This makes note of the situation and separates `version` under `Versioning` and `versionTag` remains as `VersionTag`.
Similarly, the `AsyncBefore` and `AsyncAfter` are also notable probably Camunda-specific properties.
There may be several more Camunda-specific properties.
Changing the behavior to remove Camunda-specific functionality is out of scope.

There is inconsistency in how properties are to be provided for connecting to services.
Some use a full URL and others have it broken up into parts.
This is either a previous design inconsistency or an inconsistency in Camunda.
The relevant `Has` interfaces are broken up to make room for fixing this at a later time.
Fixing the inconsistency is considered out of scope.

The password is broken out because ideally the password will not be specified all of the time.
This password should further be obfuscated.
The UI should not receive the password.
More work is needed to do this properly and that work is considered out of scope.
The only focus on the password is to separate it into its own `Has` interface.

Similarly, the username property is broken out.
It may or may not be move into a combined property as I suggested in the comments or remain in its own separate `Has` interface for privacy and security considerations.
Deciding on this is out of scope.

This commit maintains the synthetic functions and keeps them on the model.
For an example see `hasCardinalityExpression()`.
This will need to be moved out into a DTO class in a subsequent commit.

This rebuilds the RAML files.
Rebuilding this revealed that some of the properties, namely `asyncBefore` and `asyncAfter`, are missing from the previous RAML files in some cases.