MODWRKFLOW-2: Add and use DTOs, mainly Workflow.
This enables DTOs where possible.
The DTOs are used when fetching the data to reduce the data loaded from the database, such as with `WorkflowOperationDto`.
This required extracting out `HasDeploymentId` from `HasWorkflowCommon`.

The Jackson handler as it is currently used cannot properly transfer an interface DTO (aka: a DTO Projection).
A concrete class is required.
For now, the Workflow Entity is directly used here.
This should be fine for the short term given that the entire Workflow is needed for these cases anyway.
Additional work is merited to better solve this.
Consider things like ModelMapper to assist with this.

The additional DTOs that should be created, like NodeDto, are not done due to an observed problem.
The `@JsonTypeInfo` and `@JsonSubTypes` using `deserializeAs` does not translate well into the interfaces that DTOs use.
This results in an invalid identifier and the Workflow cannot be activated.
This requires more research and work to determine a valid strategy to preserve the `getIdentifier()` needs but with an interface.