Improve handling of errors when communicating with Camunda.
Catch the erros so that they can reported more properly.
This returns the actual error message in the response so that the actual error messages reported by Camunda appear in the response when using say `fw-cli activate`.

The original errors are still reported in the log and the error is passed to the exceptions service.

This change results in a response such as this:
```
failed put http://127.0.0.1:9001/workflows/9908e4c0-6c48-4b2f-a487-a0398f74549f/activate {}
{
  errors: [
    {
      message: `Failed to activate workflow: 403 Forbidden: "Access for user 'diku_admin' (aa572e0c-ca18-5024-8a78-9e1aaabffc5a) requires permission: camunda.workflow-engine.workflows.activate"!`,
      type: 'WorkflowEngineServiceException',
      code: '500 INTERNAL_SERVER_ERROR',
      parameters: []
    }
  ],
  total_records: 1
}
```