MODWRKFLOW-8: Upgrade to use Spring-Boot 3.
This makes no attempts to fix problems introduced by dependencies.
As a result, the RAML files do not get properly built.

This new dependency regression is considered acceptable because the problematic dependencies are going away as of MODWRKFLOW-10.
This change set should be in place before introducing changes from MODWRKFLOW-10.