MODWRKFLOW-6: Improve code quality according to Sonarcloud.
Change the behavior of `@Autowired` on variables and move them into a constructor as per Sonarcloud `java:S6813`.
Doing this introduces problems and maintainability issues, so beware.

Remove unnecessary braces and return statement as per Sonarcloud `java:S1602`.

Replace logic in `AbstractCqlService` and implementing classes to work-around Sonarcloud `java:S2201` lack of understanding.
Improve while at it, simplifying the design.

Swap the `abstract` on methods in what I would call bad practice to appease Sonarcloud `java:S1124`.

Remove the extra logging to appease Sonarcloud `java:S2139`.
This could be a problem as we want to see the logs but also want to pass the exception forward.
The exception potentially may be later handled in such a way that it does not log the exception.
We lose some maintainability and runtime debugging capabilities with following this opinion by Sonarcloud.

Remove unused imports.

Add missing `private` keyword on variable.