Remove Autowired constructor because it breaks constructors that are not Autowired.
The `@Autowired` on the constructor breaks behavior with the existing non-`@Autowired` constructor.
The non-`@Autowired` constructor is instantiating the `RestTemplate`.
Adding the `RestTemplate` to the `@Autowired` constructor would result in unsafe behavior where there are two allocations on `RestTemplate`.
This produces potential undefined behavior such as constructor race conditions.
This would likely cause obscure problems.

Do not use an `@Autowired` constructor in this specific case.