MODWRKFLOW-19: Allow for safe deletion of workflows.
This is designed to call the deactivate in mod-camunda on the expectation that mod-camunda is properly deactivating (setting cascade to true on deactivate/undeploy).

Disable the spring-data-rest end point for the `DELETE` on a Workflow.

Add a new dedicated end point for Workflow that performs the `DELETE`.
- This is done via a custom function in the `WorkflowEngineService`.
- An HTTP 204 must be explicitly returned because this end point does not return content.
- The mod-camunda end point is also expected to not throw an error when deactivating an already deactivated workflow.
  - This requires changes in mod-camunda.
- Fix potential logic problems in `sendWorkflowRequest` in regards to when `responseWorkflow` is NULL.
  - The `responseWorkflow` could be NULL on certain valid responses when status is not HTTP 200 (OK).

Perform "code smell" clean up by moving the `Response body: {}` message to a class constant.