MODWRKFLOW-18: Add components model unit tests.
The `spring-test` dependency is now added to the components pom.

The `maven-surefire-plugin` is now enabled for the components so that tests are actually run.

The unit tests exposed that several of the models are missing the `HasAsync` interface.
This adds the `HasAsync` to the missing models and this change should not break anything.

Add all of the components model unit tests.