MODWRKFLOW-22: Document the FWZ format.
This documents the **FWZ** by documenting the format as seen in the fw-registry project.

There are some parts of the code that the documentation is omitted due to not enough information or the use case is unclear.
One such example is the `workflow` property for the `DirectoryTask`.

This is the initial documentation.
There may be additional changes, clarifications, or improvements once the work begins on the MODWRKFLOW-21 issue.