MODWRKFLOW-22: Add import end point that accepts FWZ format.
This implements a custom mime-type detector that avoids bringing in additional dependencies (`CompressFileMagic.java`).
Use a miime-type detection based on Wikipedia mime-types database.

Bring in Apache Commons Compress dependency to perform the archive extraction and manipulation.

Rename the `TaskRepo` to `NodeRepo` so as to be less confusing.
This fixing an incomplete refactor from a long time ago.

This creates the new end point: `/import`.
On success an HTTP 201 with the created Workflow is returned.
The location of the workflow is returned in the HTTP header part of the HTTP 201 response.

New exceptions specific to the import process are added.

An `ExtractedWorkflow` class is added that is intended to reduce the number of parameters passed to a function.
Out of convenience, several related constants are provided in this class.

A new repo method `existsById` is provided out of necessity.
Not having this and calling the method without explicitly being defined resulted in errors.
The `@RestResource(exported = false)` is necessary to prevent this method from potentially being exposed via the Spring Data Rest end points.

The `.tar.gz`, `.tar.bz2`, and `.zip` formats are implemented and working.
The current design requires that the archives to not be stored within a sub-directory.
The `workflow.json`, the `fwz.json`, and related files must be at the top-level of the archive.
Adding sub-directory support for a single directory is considered out of scope for this issue.

This performs an in-memory processing of the given file on the expectation that the workflow file size is reasonably small.
Improving this to detect file size and use other methods is considered out of scope.

This is designed around the expectation that the `code` values are to be stringified for proper escaping of the code.

The files are expected to be in UTF-8 encoding.

The JSON files are automatically ordered on import based on dependencies of any particular JSON file in a top-down manner via a cache in memory using the `ExtractedWorkflow` class.