MODWRKFLOW-22: Add more unit tests, but WorkflowImportServiceTest is incomplete.
I ran out of time and did not finish the `WorkflowImportServiceTest` tests.
The tests in `WorkflowImportServiceTest` are currently disabled due to not working just yet.
I've created some example `.fwz` files using `.tar.gz`.
The tests are currently failing on it not being recognized as a **GZIP** format.
This might be because of how I am loading the file.
The files are in **GZIP** format.