MODWRKFLOW-28: Add missing or fix incorrect module descriptor end points.
Sort the end points alphabetically so that I can more easily scan the files and find mistakes.

There appears to be configuration for the Spring Data Rest end points (which may or may not be made available in the future).

There are not end points for the explicit ones, such as `/workflows/{id}/delete`.
The explicit end point for deleting a workflow is now exposed via the module descriptor.

The `/workflows/{id}/start` end point is also missing and is now exposed via the module descriptor.
The permission is assigned to the `workflow.workflows.item.startTrigger`.
There may or may not be a need to have a separate permission called `workflow.workflows.item.start`.
Such a permissions change is considered out of scope.