MODWRKFLOW-31: First pass implementation of the InputTask.
There is a large amount of inconsistency within the backend code for the Entity structure.
This makes it very confusing on which way to implement this.

I did a lot of testing and found that using `Boolean` type over `boolean` with `@PrePersist` allows for handling non-existent data from passed JSON.
A column default is now specified to further assist in ensure a consistent state in the database.
This prevents build problems and helps ensure a consistent database state.

I observed problems with building an example input task workflow and discovered that the `EmbeddedVariable` variable called `outputVariable` to be the problem.
I changed the `EmbeddedVariable` to be embedded as a JSON string rather than using the `@Embedded`.
This also prevents some confusing column overlap problems.
The `outputVariable` is optional but when it is specified, then most of its variables are required.

One of the errors I encounted is `asJson` not being defined in the newly created `InputTask`.
The example `InputTask` Workflow that I created is not using an `EmbeddedVariable`.
Making the structural change to the `outputVariable` as described above solved this problem.

The database booleans should be either `true` or `false` and not one of `true`, `false`, or `null`.

I wanted to avoid reproducing the same code so I made `AbstractDelegateTaskNode` to reduce repitition.
I updated all of the existing models to make the code consistent with this observed changes.

The existing models now have more consistent assignnment of defaults.

I have tested and confirmed that all known Workflows build under this change set.