MODWRKFLOW-31: Switch to Task and re-generate OpenAPI files.
Most of the models are a `DelegateTask`.
Two of the existing ones are note and accidentally got converted:
- `ReceiveTask`
- `ScriptTask`

These have been converted back to a `Task` rather than being a `DelegateTask`.

The first attempt at this design is to try and make the `InputTask` a `Task` rather than a `DelegateTask`.
Make the `InputTask` an `InputTask`.

Re-genereate the OpenAPI files.